python_client/requirements.txt
# Requirements for the Python client
requests>=2.25.0
numpy>=1.21.0
pandas>=1.3.0
gym>=0.21.0
gymnasium>=0.26.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=3.19.0
stable-baselines3[extra]>=2.0.0
matplotlib>=3.5.0
tensorboard>=2.8.0

# Optional: For advanced RL training
stable-baselines3>=1.0.0
torch>=1.10.0
tensorboard>=2.0.0

# Optional: For visualization
matplotlib>=3.3.0
seaborn>=0.11.0
